Subset the provided datamart data to just the latest snapshot per model.

If there is just one snapshot, nothing will change. It works for both model and predictor tables. If there is no snapshottime field, it will not do anything.

latestSnapshotsOnly(dt)

Arguments

dt
The data.table with the datamart data.

Value

A data.table with just the latest snapshots per model.

Examples

latestSnapshotsOnly(admdatamart_binning)

#>                                    ModelID RulesetName
#>    1: bd70a915-697a-5d43-ab2c-53b0557c85a0          NA
#>    2: bd70a915-697a-5d43-ab2c-53b0557c85a0          NA
#>    3: bd70a915-697a-5d43-ab2c-53b0557c85a0          NA
#>    4: bd70a915-697a-5d43-ab2c-53b0557c85a0          NA
#>    5: bd70a915-697a-5d43-ab2c-53b0557c85a0          NA
#>   ---                                                 
#> 4572: 277a2c48-8888-5b71-911e-443d52c9b50f          NA
#> 4573: 277a2c48-8888-5b71-911e-443d52c9b50f          NA
#> 4574: 277a2c48-8888-5b71-911e-443d52c9b50f          NA
#> 4575: 277a2c48-8888-5b71-911e-443d52c9b50f          NA
#> 4576: 277a2c48-8888-5b71-911e-443d52c9b50f          NA
#>                   PredictorName              Contents     Type      BinType
#>    1:           Customer.Prefix       No. symbols = 5 symbolic EQUIBEHAVIOR
#>    2: Customer.IsCustomerActive       No. symbols = 2 symbolic EQUIBEHAVIOR
#>    3:     Customer.HealthMatter       No. symbols = 1 symbolic      MISSING
#>    4:              Customer.SSN       No. symbols = 1 symbolic EQUIBEHAVIOR
#>    5:        Customer.pyCountry       No. symbols = 3 symbolic EQUIBEHAVIOR
#>   ---                                                                      
#> 4572:                Classifier Range : [-1.87, 0.15]  numeric EQUIBEHAVIOR
#> 4573:                Classifier Range : [-1.87, 0.15]  numeric EQUIBEHAVIOR
#> 4574:                Classifier Range : [-1.87, 0.15]  numeric EQUIBEHAVIOR
#> 4575:                Classifier Range : [-1.87, 0.15]  numeric EQUIBEHAVIOR
#> 4576:                Classifier Range : [-1.87, 0.15]  numeric EQUIBEHAVIOR
#>            BinSymbol BinLowerBound BinUpperBound  EntryType Extension
#>    1:    NON-MISSING            NA            NA   Inactive        NA
#>    2:    NON-MISSING            NA            NA   Inactive        NA
#>    3:        MISSING            NA            NA   Inactive        NA
#>    4:    NON-MISSING            NA            NA   Inactive        NA
#>    5:    NON-MISSING            NA            NA   Inactive        NA
#>   ---                                                                
#> 4572: [0.065, 0.085>         0.065         0.085 Classifier        NA
#> 4573:  [0.085, 0.09>         0.085         0.090 Classifier        NA
#> 4574:  [0.09, 0.125>         0.090         0.125 Classifier        NA
#> 4575: [0.125, 0.995>         0.125         0.995 Classifier        NA
#> 4576:        >=0.995         0.995         1.000 Classifier        NA
#>       CorrelationPredictor        SnapshotTime Performance Positives Negatives
#>    1:                   NA 2021-06-01 15:23:26    0.500000        12      5805
#>    2:                   NA 2021-06-01 15:23:26    0.500000        12      5805
#>    3:                   NA 2021-06-01 15:23:26    0.500000        12      5805
#>    4:                   NA 2021-06-01 15:23:26    0.500000         0         0
#>    5:                   NA 2021-06-01 15:23:26    0.500000        12      5805
#>   ---                                                                         
#> 4572:                   NA 2021-06-01 15:23:26    0.773211       536      4440
#> 4573:                   NA 2021-06-01 15:23:26    0.773211       536      4440
#> 4574:                   NA 2021-06-01 15:23:26    0.773211       536      4440
#> 4575:                   NA 2021-06-01 15:23:26    0.773211       536      4440
#> 4576:                   NA 2021-06-01 15:23:26    0.773211       536      4440
#>       TotalBins ResponseCount RelativePositives RelativeNegatives
#>    1:         1          5817                NA                NA
#>    2:         1          5817                NA                NA
#>    3:         1          5817                NA                NA
#>    4:         1             0                NA                NA
#>    5:         1          5817                NA                NA
#>   ---                                                            
#> 4572:        22          4976                NA                NA
#> 4573:        22          4976                NA                NA
#> 4574:        22          4976                NA                NA
#> 4575:        22          4976                NA                NA
#> 4576:        22          4976                NA                NA
#>       RelativeResponseCount BinNegatives BinPositives BinNegativesPercentage
#>    1:                    NA         5805           12                 100.00
#>    2:                    NA         5805           12                 100.00
#>    3:                    NA         5805           12                 100.00
#>    4:                    NA            0            0                   0.00
#>    5:                    NA         5805           12                 100.00
#>   ---                                                                       
#> 4572:                    NA           10           13                   0.23
#> 4573:                    NA            1            2                   0.02
#> 4574:                    NA            6           15                   0.14
#> 4575:                    NA            1            4                   0.02
#> 4576:                    NA            0            1                   0.00
#>       BinPositivesPercentage RelativeBinPositives RelativeBinNegatives
#>    1:                 100.00                   NA                   NA
#>    2:                 100.00                   NA                   NA
#>    3:                 100.00                   NA                   NA
#>    4:                   0.00                   NA                   NA
#>    5:                 100.00                   NA                   NA
#>   ---                                                                 
#> 4572:                   2.43                   NA                   NA
#> 4573:                   0.37                   NA                   NA
#> 4574:                   2.80                   NA                   NA
#> 4575:                   0.75                   NA                   NA
#> 4576:                   0.19                   NA                   NA
#>       BinResponseCount RelativeBinResponseCount BinResponseCountPercentage
#>    1:             5817                       NA                     100.00
#>    2:             5817                       NA                     100.00
#>    3:             5817                       NA                     100.00
#>    4:                0                       NA                       0.00
#>    5:             5817                       NA                     100.00
#>   ---                                                                     
#> 4572:               23                       NA                       0.46
#> 4573:                3                       NA                       0.06
#> 4574:               21                       NA                       0.42
#> 4575:                5                       NA                       0.10
#> 4576:                1                       NA                       0.02
#>       BinIndex     Lift   ZRatio GroupIndex PerformanceError FeatureImportance
#>    1:        1 1.000000 0.000000          0                0                 0
#>    2:        1 1.000000 0.000000          0                0                 0
#>    3:        1 1.000000 0.000000          0                0                 0
#>    4:        1 1.000000 0.000000          1                0                 0
#>    5:        1 1.000000 0.000000          0                0                 0
#>   ---                                                                         
#> 4572:       18 5.247242 3.292316          0                0                 0
#> 4573:       19 6.189055 1.326496          0                0                 0
#> 4574:       20 6.631130 3.727502          0                0                 0
#> 4575:       21 7.426866 1.943355          0                0                 0
#> 4576:       22 9.283582 1.000934          0                0                 0