Performance of predictors per proposition

Creates a visual grid of predictor names vs propositions (configurable) with colors indicating the strength of the predictor for that proposition.

plotADMPredictorPerformanceMatrix(
  predictordata,
  modeldata,
  aggregation = intersect(c("Issue", "Group", "Name", "Treatment"), names(modeldata)),
  facets = "ConfigurationName",
  limit = .Machine$integer.max
)

Arguments

predictordata
ADM Datamart predictor data. Binning data not used and can be left out.

modeldata
ADM Datamart model data.

aggregation
Vector of field names to aggregate by. Defaults to "Issue","Group","Name","Treatment" (if present).

facets
Vector of additional fields for faceting. Defaults to "ConfigurationName".

limit
Optional limit of number of predictors. Useful if there are many predictors to keep the plot readable.

Value

A ggplot object that can either be printed directly, or to which additional decoration (e.g. coloring or labels) can be added first.

Examples

if (FALSE) {
    plt <- plotADMPredictorPerformanceMatrix(modelPredictorBins, mdls, limit=50) +
    theme(axis.text.y = element_text(size=8),
    axis.text.x = element_text(size=8, angle = 45, hjust = 1),
    strip.text = element_text(size=8))
    print(plt)
}