Generates PMML files from ADM models.

Generates PMML files from the ADM models in either the ADM data mart, directly from the database or from dataset exports, or from the (internal) "Factory" tables. The generated PMML contains all the individual models for each model partition inside a big ensemble that selects between them based on context keys.

adm2pmml(
  dmModels = NULL,
  dmPredictors = NULL,
  dbConn = NULL,
  forceUseDM = FALSE,
  appliesToFilter = NULL,
  ruleNameFilter = NULL,
  destDir = ".",
  tmpDir = NULL,
  verbose = (!is.null(tmpDir))
)

Arguments

dmModels
Data frame (or anything else that converts to data.table) with the ADM datamart model data

dmPredictors
Data frame (or anything else that converts to data.table) with the ADM datamart predictor data. Required if dmModels is given. If dmPredictors is given but dmModels is not, the assumption will be made that all the predictor binning belongs to a single model.

dbConn
Database connection. If given, the data will be extracted from the internal ADM Factory table unless the forceUseDM flag is set.

forceUseDM
Only in combination with the dbConn flag. If TRUE, will use the ADM datamart tables. If FALSE (the default) will use the ADM Factory tables.

appliesToFilter
Optional filter for the applies to class of the models as a regexp passed on to grepl with ignore.case = TRUE, perl = TRUE, fixed = FALSE, useBytes = FALSE.

ruleNameFilter
Optional filter for the rule name of the models as a regexp passed on to grepl with ignore.case = TRUE, perl = TRUE, fixed = FALSE, useBytes = FALSE.

destDir
Directory to create the PMML files in. Defaults to the current directory.

tmpDir
Directory for temporary files. Defaults to NULL indicating no temp files will be left behind.

verbose
Whether to print the steps it's performing. Defaults to TRUE if the tmpDir is set.

Value

Creates the PMML file(s) in the given destination and returns a list of model ID's per file.

Exporting a specific version

When exporting from the datamart, only the models from the most recent model snapshot will be exported, and for each of these, only the most recent of the predictor snapshots of that model will be used. To export specific snapshots subset the data before calling this method.

Examples

adm2pmml(verbose=TRUE)

#> [1] "No models available matching the filter criteria"

#> list()

if (FALSE) {
  models <- readDSExport("Data-Decision-ADM-ModelSnapshot_AllModelSnapshots", "~/Downloads")
  preds <- readDSExport("Data-Decision-ADM-PredictorBinningSnapshot_AllPredictorSnapshots",
                        "~/Downloads")

  adm2pmml(models, preds, verbose=TRUE)
}